+++ Sources/TuistCache/Cache/Responses/CloudVerifyUploadResponse.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+struct CloudVerifyUploadResponse: Decodable {
+    let uploadedSize: Int
+}