+++ Sources/TuistCacheTesting/Cache/TestData/LabVerifyUploadResponse+TestData.swift
@@ -1,7 +1,0 @@
-@testable import TuistCache
-
-extension LabVerifyUploadResponse {
-    public static func test(uploadedSize: Int = 0) -> LabVerifyUploadResponse {
-        LabVerifyUploadResponse(uploadedSize: uploadedSize)
-    }
-}