+++ Sources/TuistKit/ProjectEditor/ProjectEditorMapper.swift
@@ -254,7 +254,7 @@ final class ProjectEditorMapper: ProjectEditorMapping {
         let projectSettings = Settings(
             base: [
                 "ONLY_ACTIVE_ARCH": "NO",
-                "EXCLUDED_ARCHS": "arm64",
+                "EXCLUDED_ARCHS": .string(excludedArchs()),
             ],
             configurations: Settings.default.configurations,
             defaultSettings: .recommended
@@ -267,6 +267,7 @@ final class ProjectEditorMapper: ProjectEditorMapping {
             name: projectName,
             organizationName: nil,
             developmentRegion: nil,
+            options: [],
             settings: projectSettings,
             filesGroup: manifestsFilesGroup,
             targets: targets,
@@ -334,7 +335,7 @@ final class ProjectEditorMapper: ProjectEditorMapping {
         let projectSettings = Settings(
             base: [
                 "ONLY_ACTIVE_ARCH": "NO",
-                "EXCLUDED_ARCHS": "arm64",
+                "EXCLUDED_ARCHS": .string(excludedArchs()),
             ],
             configurations: Settings.default.configurations,
             defaultSettings: .recommended
@@ -347,6 +348,7 @@ final class ProjectEditorMapper: ProjectEditorMapping {
             name: projectName,
             organizationName: nil,
             developmentRegion: nil,
+            options: [],
             settings: projectSettings,
             filesGroup: pluginsFilesGroup,
             targets: pluginTargets,
@@ -424,7 +426,15 @@ final class ProjectEditorMapper: ProjectEditorMapping {
 
     private func targetBaseSettings(for includes: [AbsolutePath], swiftVersion: String) -> SettingsDictionary {
         let includePaths = includes
-            .map(\.parentDirectory.pathString)
+            .flatMap { path in
+                // In development, the .swiftmodule is generated in a directory up from the directory of the framework.
+                // /path/to/derived/tuist-xyz/
+                //    PackageFrameworks/
+                //      ProjectDescription.framework
+                //    ProjectDescription.swiftmodule
+                // Because of that we need to expose the parent directory too in SWIFT_INCLUDE_PATHS
+                [path.parentDirectory.pathString, path.parentDirectory.parentDirectory.pathString]
+            }
             .map { "\"\($0)\"" }
         return [
             "FRAMEWORK_SEARCH_PATHS": .array(includePaths),
@@ -432,5 +442,15 @@ final class ProjectEditorMapper: ProjectEditorMapping {
             "SWIFT_INCLUDE_PATHS": .array(includePaths),
             "SWIFT_VERSION": .string(swiftVersion),
         ]
+    }
+
+    private func excludedArchs() -> String {
+        let architecture = DeveloperEnvironment.shared.architecture
+        switch architecture {
+        case .arm64:
+            return MacArchitecture.x8664.rawValue
+        case .x8664:
+            return MacArchitecture.arm64.rawValue
+        }
     }
 }