+++ Tests/TuistGeneratorTests/Generator/ProjectGroupsTests.swift
@@ -28,6 +28,7 @@ final class ProjectGroupsTests: XCTestCase {
             name: "Project",
             organizationName: nil,
             developmentRegion: nil,
+            options: [],
             settings: .default,
             filesGroup: .group(name: "Project"),
             targets: [
@@ -169,5 +170,40 @@ final class ProjectGroupsTests: XCTestCase {
 
     func test_projectGroupsError_type() {
         XCTAssertEqual(ProjectGroupsError.missingGroup("abc").type, .bug)
+    }
+
+    func test_generate_with_text_settings() {
+        // Given
+        let textSettings = TextSettings.test()
+        let project = Project.test(options: [.textSettings(textSettings)])
+
+        // When
+        let main = ProjectGroups.generate(
+            project: project,
+            pbxproj: pbxproj
+        ).sortedMain
+
+        // Then
+        XCTAssertEqual(main.usesTabs, textSettings.usesTabs)
+        XCTAssertEqual(main.indentWidth, textSettings.indentWidth)
+        XCTAssertEqual(main.tabWidth, textSettings.tabWidth)
+        XCTAssertEqual(main.wrapsLines, textSettings.wrapsLines)
+    }
+
+    func test_generate_without_text_settings() {
+        // Given
+        let project = Project.test(options: [])
+
+        // When
+        let main = ProjectGroups.generate(
+            project: project,
+            pbxproj: pbxproj
+        ).sortedMain
+
+        // Then
+        XCTAssertNil(main.usesTabs)
+        XCTAssertNil(main.indentWidth)
+        XCTAssertNil(main.tabWidth)
+        XCTAssertNil(main.wrapsLines)
     }
 }