+++ Tests/TuistIntegrationTests/Generator/MultipleConfigurationsIntegrationTests.swift
@@ -351,7 +351,7 @@ final class MultipleConfigurationsIntegrationTests: TuistUnitTestCase {
     }
 
     private func createConfig() -> Config {
-        Config(compatibleXcodeVersions: .all, lab: nil, cache: .default, swiftVersion: nil, plugins: [], generationOptions: [], path: nil)
+        Config(compatibleXcodeVersions: .all, cloud: nil, cache: .default, swiftVersion: nil, plugins: [], generationOptions: [], path: nil)
     }
 
     private func createWorkspace(path: AbsolutePath, projects: [String]) throws -> Workspace {
@@ -371,6 +371,7 @@ final class MultipleConfigurationsIntegrationTests: TuistUnitTestCase {
             name: "App",
             organizationName: nil,
             developmentRegion: nil,
+            options: [],
             settings: settings,
             filesGroup: .group(name: "Project"),
             targets: targets,