+++ Tests/TuistSupportTests/Utils/GitHubClientIntegrationTests.swift
@@ -1,0 +1,7 @@
+import Foundation
+import XCTest
+
+@testable import TuistSupport
+@testable import TuistSupportTesting
+
+final class GitHubClientIntegrationTests: TuistTestCase {}