+++ Tests/TuistSupportTests/Utils/GitHubClientTests.swift
@@ -32,7 +32,7 @@ final class GitHubClientTests: TuistUnitTestCase {
 
         // When
         let expectation = XCTestExpectation(description: "GitHubClient deferred when token")
-        _ = subject.deferred(resource: HTTPResource<Void, Never>.noop())
+        _ = subject.dispatch(resource: HTTPResource<Void, Never>.noop())
             .sink(
                 receiveCompletion: { _ in expectation.fulfill() },
                 receiveValue: { _ in }
@@ -54,7 +54,7 @@ final class GitHubClientTests: TuistUnitTestCase {
 
         // When
         let expectation = XCTestExpectation(description: "GitHubClient deferred when token")
-        _ = subject.deferred(resource: HTTPResource<Void, Never>.noop())
+        _ = subject.dispatch(resource: HTTPResource<Void, Never>.noop())
             .sink(
                 receiveCompletion: { _ in expectation.fulfill() },
                 receiveValue: { _ in }
@@ -76,7 +76,7 @@ final class GitHubClientTests: TuistUnitTestCase {
 
         // When
         let expectation = XCTestExpectation(description: "GitHubClient deferred when token")
-        _ = subject.deferred(resource: HTTPResource<Void, Never>.noop())
+        _ = subject.dispatch(resource: HTTPResource<Void, Never>.noop())
             .sink(
                 receiveCompletion: { _ in expectation.fulfill() },
                 receiveValue: { _ in }
@@ -88,5 +88,20 @@ final class GitHubClientTests: TuistUnitTestCase {
         let headers = try XCTUnwrap(request.allHTTPHeaderFields)
         XCTAssertNil(headers["Authorization"])
         XCTAssertEqual(headers["Accept"], "application/vnd.github.v3+json")
+    }
+
+    func test_something() throws {
+        let expectation = XCTestExpectation(description: "GitHubClient deferred when token")
+        var release: GitHubRelease?
+        let client = GitHubClient()
+        _ = client.dispatch(resource: GitHubRelease.latest(repositoryFullName: "tuist/tuist"))
+            .sink { error in
+                print(error)
+                expectation.fulfill()
+            } receiveValue: { response in
+                print(response)
+                release = response.object
+            }
+        wait(for: [expectation], timeout: 10.0)
     }
 }