+++ projects/tuist/features/step_definitions/shared/developer_environment.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require "fileutils"
+
+Then(/I install the Workflow extensions SDK/) do
+  sdk_pkg_path = File.expand_path("../../resources/WorkflowExtensionsSDK.pkg", __dir__)
+  unless File.exist?("/Library/Developer/SDKs/WorkflowExtensionSDK.sdk")
+    system("sudo", "installer", "-package", sdk_pkg_path, "-target", "/")
+  end
+end